Implement search feature
This is a search example, through Twitter-like hashtags.
We created a separate query and service to decouple our code and not
clump a lot of stuff into the main models (user and domain-specific main
model, Shout).
We are also using Apache Solr as our search provider, since it's much
better to use an external tool for this instead of rolling our own
solution, even more so considering that we could have millions of Shouts
in the future.